Create SettingsSubInterface.res
This file is a bit of a work in progress, there are additional control included but I left them out since I can't test them. One glaring issue that I was not able to solve is the last check box, the text isn't aligned properly for some reason. Whilst it is an extremely minor issue that doesn't affect UX in anyway, my eyes cannot unsee it.